+++ chevah/compat/nt_filesystem.py
@@ -147,26 +147,41 @@ class NTFilesystem(PosixFilesystemBase):
         return unicode(result)
 
     def getSegmentsFromRealPath(self, path):
-        '''See `ILocalFilesystem`.'''
+        """
+        See `ILocalFilesystem`.
+        """
         segments = []
 
         if path is None or path == u'':
             return segments
 
-        path = os.path.abspath(path)
-        drive, root_tail = os.path.splitdrive(path)
-        if drive == u'':
-            segments = [u'c']
-        else:
-            segments = [drive.strip(u':')]
-
         head = True
-        tail = root_tail
+        path = os.path.abspath(path)
+
+        if self._avatar.lock_in_home_folder:
+            # Locked filesystem have no drive.
+            tail = path[len(self._getRootPath()):]
+            drive = ''
+        else:
+            # For unlocked filesystem, we use 'c' as default drive.
+            drive, root_tail = os.path.splitdrive(path)
+            if not drive:
+                drive = u'c'
+            else:
+                drive = drive.strip(u':')
+            tail = root_tail
+
         while head not in [u'\\', u'']:
             head, tail = os.path.split(tail)
-            if tail != '':
-                segments.insert(1, tail)
+            if tail == '':
+                break
+            segments.insert(0, tail)
             tail = head
+
+        # Prepend drive at the end, due to the way os.path.split works.
+        if drive:
+            segments.insert(0, drive)
+
         return segments
 
     def readLink(self, segments):